ref: código refatorado para utilizar fragments ao inves de acvitivities. Camera Activity substituido por Camera Fragment
