+++ app/src/main/res/layout/camera_fragment.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/cameraActivity"
+    android:id="@+id/cameraFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
     <FrameLayout
-
+        android:id="@+id/cameraFrameLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
@@ -39,5 +40,4 @@
         app:layout_constraintEnd_toEndOf="parent"
         android:layout_marginBottom="48dp"
         android:background="@drawable/cameraswitch"/>
-
 </androidx.constraintlayout.widget.ConstraintLayout>
