+++ app/src/main/res/layout/usb_camera_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CameraActivity">
+    tools:context=".MainActivity">
 
     <com.serenegiant.widget.AspectRatioSurfaceView
         android:id="@+id/svCameraViewMain"
